completed- initalization and refactor of app.py
