+++ frontend-react-js/src/pages/NotificationsFeedPage.js
@@ -6,9 +6,8 @@ import DesktopSidebar     from '../components/DesktopSidebar';
 import ActivityFeed from '../components/ActivityFeed';
 import ActivityForm from '../components/ActivityForm';
 import ReplyForm from '../components/ReplyForm';
+import {checkAuth, getAccessToken} from 'lib/CheckAuth';
 
-// [TODO] Authenication
-import Cookies from 'js-cookie'
 
 export default function NotificationsFeedPage() {
   const [activities, setActivities] = React.useState([]);
@@ -21,7 +20,13 @@ export default function NotificationsFeedPage() {
   const loadData = async () => {
     try {
       const backend_url = `${process.env.REACT_APP_BACKEND_URL}/api/activities/notifications`
+      await getAccessToken()
+      const access_token = localStorage.getItem("access_token")
       const res = await fetch(backend_url, {
+        headers: {
+          Authorization: `Bearer ${access_token}`
+        },
+
         method: "GET"
       });
       let resJson = await res.json();
@@ -35,24 +40,15 @@ export default function NotificationsFeedPage() {
     }
   };
 
-  const checkAuth = async () => {
-    console.log('checkAuth')
-    // [TODO] Authenication
-    if (Cookies.get('user.logged_in')) {
-      setUser({
-        display_name: Cookies.get('user.name'),
-        handle: Cookies.get('user.username')
-      })
-    }
-  };
-
+ 
   React.useEffect(()=>{
     //prevents double call
     if (dataFetchedRef.current) return;
     dataFetchedRef.current = true;
 
     loadData();
-    checkAuth();
+    checkAuth(setUser);
+
   }, [])
 
   return (
